tests: bluetooth: classic: Refactor test suites
Multiple Bluetooth Classic test suites contained duplicate
implementations of common functions including `device_power_on()`,
`wait_for_shell_response()`, and `send_cmd_to_iut()`.

Creates a new utility module under `tests/bluetooth/classic/utility`.

Consolidates common functions into `utility/{common,gap,l2cap}.py`.

Updates all test suites to import from the shared utility modules.

Removes duplicate function definitions across test files.

Maintains existing functionality while improving code maintainability.
