+++ tests/bluetooth/classic/utility/__init__.py
@@ -1,0 +1,3 @@
+# Copyright 2025 NXP
+#
+# SPDX-License-Identifier: Apache-2.0
